Refractor replacements into own file
